Modularized LibraryItemScreen
