+++ features/libraries/api/src/commonMain/kotlin/app/campfire/libraries/api/screen/LibraryItemScreen.kt
@@ -1,0 +1,10 @@
+package app.campfire.libraries.api.screen
+
+import app.campfire.common.screens.DetailScreen
+import app.campfire.core.model.LibraryItemId
+import app.campfire.core.parcelize.Parcelize
+
+@Parcelize
+data class LibraryItemScreen(
+  val libraryItemId: LibraryItemId,
+) : DetailScreen(name = "LibraryItem()")
