+++ features/sessions/ui/src/commonMain/kotlin/app/campfire/sessions/ui/ExpandedPlaybackBar.kt
@@ -84,7 +84,7 @@ import app.campfire.common.compose.layout.isSupportingPaneEnabled
 import app.campfire.common.compose.theme.PaytoneOneFontFamily
 import app.campfire.common.compose.widgets.CoverImage
 import app.campfire.common.compose.widgets.CoverImageSize
-import app.campfire.common.screens.LibraryItemScreen
+import app.campfire.libraries.api.screen.LibraryItemScreen
 import app.campfire.core.extensions.fluentIf
 import app.campfire.core.model.Bookmark
 import app.campfire.core.model.Chapter
